+++ poptorch/source/PoptorchSymbols.hpp
@@ -79,6 +79,10 @@ extern c10::Symbol suppress_autocast;
 extern c10::Symbol restore_autocast;
 extern c10::Symbol autocast;
 
+// In order to allow a paramater/buffer to be updated in place, the only
+// gauranteed inplace op by PopART, use update_param_inplace.
+extern c10::Symbol update_param_inplace;
+
 // Casting is done before passing the input to the IPU: the op is used so that
 // so that that input types match those received from pytorch but that the input
 // types to later ops have the correct type.